+++ src/pages/paint/paintMenuMisc.tsx
@@ -1,4 +1,5 @@
 // this depends on PainMenu.css
+import { Color } from 'csstype';
 import * as React from 'react';
 
 export function PaintMenuContent (props: any) {
@@ -41,5 +42,98 @@ export function PaintMenuFooterButton (props: any) {
       className={`PaintMenu-footerButton ${className}`}
       {...rest}
       >{children}</button>
+  );
+}
+
+interface IStrokeWidthButtonProps {
+  width: number;
+  onClick: (width: number) => void;
+}
+function StrokeWidth (props: IStrokeWidthButtonProps) {
+  const path = `
+    M 40 10
+    L 10 40
+  `;
+  const onClick = () => props.onClick(props.width);
+  return (
+    <span className="StrokeWidth">
+      <button className="StrokeWidth-button"
+        onClick={onClick}
+        >
+          <svg width="50" height="50">
+            <path
+              stroke="#333"
+              strokeWidth={props.width}
+              d={path}
+              />
+          </svg>
+        </button>
+    </span>
+  );
+}
+
+interface IStrokeWidthsButtonProps {
+  strokeWidths: number[];
+  onChange: (width: number) => void;
+}
+export function StrokeWidths (props: IStrokeWidthsButtonProps) {
+  const onClick = (width: number) => {
+    props.onChange(width);
+  };
+
+  const buttons = props.strokeWidths.map((width: number) => {
+    return (
+      <StrokeWidth
+        key={width}
+        width={width}
+        onClick={onClick}
+        />
+    );
+  });
+
+  return (
+    <div className="StrokeWidths">
+      {buttons}
+    </div>
+  );
+}
+
+interface IColorButtonProps {
+  color: Color;
+  onClick: (color: Color) => void;
+}
+function ColorButton (props: IColorButtonProps) {
+  const onClick = () => props.onClick(props.color);
+  return (
+    <span className="ColorButton">
+      <button className="ColorButton-button"
+        style={{ backgroundColor: props.color }}
+        onClick={onClick}
+        >{props.color}</button>
+    </span>
+  );
+}
+
+interface IColorsProps {
+  colors: string[];
+  onChange: (color: Color) => void;
+}
+export function Colors (props: IColorsProps) {
+  const onClick = (color: Color) => {
+    props.onChange(color);
+  };
+
+  const buttons = props.colors.map((color) => (
+      <ColorButton
+        key={color}
+        color={color}
+        onClick={onClick}
+        />
+  ));
+
+  return (
+    <div className="PaintMenu-colors">
+      {buttons}
+    </div>
   );
 }
